part3/hbnb/requirements.txt
aniso8601==10.0.1
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
PyJWT==2.11.0
referencing==0.37.0
rpds-py==0.30.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
